# sys-service/api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Aevons 系统服务接口文档，提供用户、角色、部门、岗位、菜单、字典、语言、配置、通知公告和终端应用等系统管理能力",
        "title": "Aevons System Service API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:10702",
    "basePath": "/api/v1/sys",
    "paths": {
...
        "/api/v1/sys/user": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "新增用户",
                "parameters": [
                    {
                        "description": "新增用户",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/user/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "查询用户列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/user/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "分页查询用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/user/{ids}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "批量删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "逗号分隔的用户ID",
                        "name": "ids",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "修改用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改用户",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/user/{id}/relations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户关联信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/user/{id}/reset-password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "重置用户密码",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "重置密码",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ResetPasswordDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/user/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新用户状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateUserDTO": {
            "type": "object",
            "required": [
                "nickname",
                "password",
                "username"
            ],
            "properties": {
                "autograph": {
                    "description": "电子签名",
                    "type": "string",
                    "maxLength": 255
                },
                "avatar": {
                    "description": "头像",
                    "type": "string",
                    "maxLength": 255
                },
                "deptPosts": {
                    "description": "部门岗位列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserDeptPostDTO"
                    }
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "maxLength": 64
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string",
                    "maxLength": 11
                },
                "nickname": {
                    "description": "呢称",
                    "type": "string",
                    "maxLength": 50
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "maxLength": 128
                },
                "roleIds": {
                    "description": "角色ID列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sex": {
                    "description": "性别（0未知 1男 2女）",
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0正常 1停用）",
                    "type": "integer"
                },
                "type": {
                    "description": "用户类型",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 32
                }
            }
        },
...
        "dto.ResetPasswordDTO": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 128
                }
            }
        },
...
        "dto.UpdateUserDTO": {
            "type": "object",
            "required": [
                "id",
                "nickname",
                "username"
            ],
            "properties": {
                "autograph": {
                    "description": "电子签名",
                    "type": "string",
                    "maxLength": 255
                },
                "avatar": {
                    "description": "头像",
                    "type": "string",
                    "maxLength": 255
                },
                "deptPosts": {
                    "description": "部门岗位列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserDeptPostDTO"
                    }
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "maxLength": 64
                },
                "id": {
                    "description": "用户编号",
                    "type": "string",
                    "example": "0"
                },
                "mobile": {
                    "description": "手机号",
                    "type": "string",
                    "maxLength": 11
                },
                "nickname": {
                    "description": "呢称",
                    "type": "string",
                    "maxLength": 50
                },
                "password": {
                    "description": "密码（为空则不修改）",
                    "type": "string",
                    "maxLength": 128
                },
                "roleIds": {
                    "description": "角色ID列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sex": {
                    "description": "性别（0未知 1男 2女）",
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0正常 1停用）",
                    "type": "integer"
                },
                "type": {
                    "description": "用户类型",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 32
                }
            }
        },
        "dto.UserDeptPostDTO": {
            "type": "object",
            "properties": {
                "deptId": {
                    "type": "string",
                    "example": "0"
                },
                "postId": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer 访问令牌，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}